Finished: #247 Enhance login call to also receive the FCM token and kept v1 for backward compatibility.
